requirements.txt
discord.py
playwright
